# VMS-7647
# VMS-7647 Complete backout
VMS-7368: Fixed server linux x64 distribution building
VMS-7368: Linux and ARM distribution building scripts fixed and improved
VMS-7368: Linux desktop client distro: keep original filtered files in 'debian/'
VMS-7368: Linux distribution building scripts: fixed Qt libs filtering - unwanted libs removed
VMS-7368: Mediaserver plugins now have to be explicitly included into distribution
VMS-7567: 265 support for DW. bug fix
VMS-7647 Add GTest PrintTo(const T&, ::std::ostream*) in NX toString
VMS-7844: Build NxTool installer with CMake
VMS-7844: Build NxTool with CMake (Review fixes)
VMS-7844: Build NxTool with CMake without precompiled headers
VMS-7852: Axis analytics driver - gcc compilation fixed
VMS-7852: Axis analytics driver disabled until gcc compilation is fixed
VMS-7852: axis amalytics driver implemented
VMS-7852: axis analytics driver
VMS-7852: axis analytics driver implemented
VMS-7854: Implement Hikvision analytics driver
VMS-7854: add eventTypeList to startFetchingMetadata
VMS-7854: add log messages
VMS-7854: add plugin to installer
VMS-7854: codestyle minor
VMS-7854: minor fixes and codestyle
VMS-7854: move classes to a separate namespace
VMS-7854: reopen connection minor improvment
VMS-7854: revert debug code
VMS-7854: review
VMS-7878: Fix local file settings previews
VMS-7906: review minor
VMS-7922: Rename some variables in mac distribution setup
VMS-7935 - Reivew fixes.
VMS-7970: Add qt.conf to windows client installer and update package
VMS-7996: bug fix
VMS-8016: Add lexical alias for USB storage type
VMS-8035: Show warning if user enables recording on USB storage
VMS-8049: Proper order of manager/handler destruction
VMS-8049: Properly set resource access on manager creation
VMS-8049: Review fixes
VMS-8067: Add default HTTP port in resource searcher if it lacks in original discovered URL
VMS-8067: Native SUNAPI discovery protocol has higher priority than UPnP
VMS-8073 Fix Server crash detected
VMS-8075: Use help 3.1 in all customization until actual packages are ready
VMS-8076: Fix invalid customization cmake files
VMS-8080: add logs
VMS-8080: improve logs.
VMS-8080: review fixes
VMS-8085 Fixed strict aliasing rule violation
VMS-8092: Crash detected in server: 8667bc467a215479b582facb513fa2eb
VMS-8099: Remove unneeded namespaces from ONVIF AbsoluteMove request
VMS-8102: Generate Qt Release Pdb Packages with CMake
VMS-8103: Mediaserver (nx1) was crashed (build 17560)
WEB-792: Some advanced settings aren't shown

<==== Mercurial data ====>
Branch: master

+++ edge_firmware/maven/bin-resources/create_arm_installer.sh
@@ -42,6 +42,8 @@ else
     LIB_INSTALL_PATH="$INSTALL_PATH/mediaserver/lib"
 fi
 
+LOG_FILE="$LOGS_DIR/create_arm_installer.log"
+
 #--------------------------------------------------------------------------------------------------
 
 createArchive() # archive dir command...
@@ -447,7 +449,7 @@ copyToolchainLibsIfNeeded()
 buildInstaller()
 {
     echo ""
-    echo "Creating installer .tar.gz"
+    echo "Creating distribution .tar.gz"
     if [ ! -z "$SYMLINK_INSTALL_PATH" ]; then
         mkdir -p "$TAR_DIR/$(dirname "$SYMLINK_INSTALL_PATH")"
         ln -s "/$INSTALL_PATH" "$TAR_DIR/$SYMLINK_INSTALL_PATH"
@@ -500,7 +502,7 @@ buildDebugSymbolsArchive()
 #--------------------------------------------------------------------------------------------------
 
 # [in] WORK_DIR
-createArmInstaller()
+buildDistribution()
 {
     local -r TAR_DIR="$WORK_DIR/tar"
     local -r INSTALL_DIR="$TAR_DIR/$INSTALL_PATH"
@@ -516,7 +518,7 @@ createArmInstaller()
         mkdir -p "$ALT_LIB_INSTALL_DIR"
     fi
 
-    echo "Creating version.txt: $VERSION"
+    echo "Generating version.txt: $VERSION"
     echo "$VERSION" >"$INSTALL_DIR/version.txt"
 
     echo "Copying build_info.txt"
@@ -547,27 +549,32 @@ createArmInstaller()
 
 #--------------------------------------------------------------------------------------------------
 
-help()
+showHelp()
 {
     echo "Options:"
     echo " --no-client: Do not pack Lite Client."
     echo " -v, --verbose: Do not redirect output to a log file."
+    echo " -k, --keep-work-dir: Do not delete work directory on success."
 }
 
 # [out] LITE_CLIENT
+# [out] KEEP_WORK_DIR
 # [out] VERBOSE
 parseArgs() # "$@"
 {
     LITE_CLIENT=1
+    KEEP_WORK_DIR=0
     VERBOSE=0
 
     local ARG
     for ARG in "$@"; do
         if [ "$ARG" = "-h" -o "$ARG" = "--help" ]; then
-            help
+            showHelp
             exit 0
         elif [ "$ARG" = "--no-client" ] ; then
             LITE_CLIENT=0
+        elif [ "$ARG" = "-k" ] || [ "$ARG" = "--keep-work-dir" ]; then
+            KEEP_WORK_DIR=1
         elif [ "$ARG" = "-v" ] || [ "$ARG" = "--verbose" ]; then
             VERBOSE=1
         fi
@@ -576,8 +583,6 @@ parseArgs() # "$@"
 
 redirectOutputToLog()
 {
-    local -r LOG_FILE="$LOGS_DIR/create_arm_installer.log"
-
     echo "  See the log in $LOG_FILE"
 
     mkdir -p "$LOGS_DIR" #< In case log dir is not created yet.
@@ -609,24 +614,31 @@ onExit()
 
 main()
 {
+    declare -i VERBOSE #< Mot local - needed by onExit().
+    local -i LITE_CLIENT
+    parseArgs "$@"
+
     trap onExit EXIT
 
     local -r WORK_DIR="$BUILD_DIR/create_arm_installer_tmp"
     rm -rf "$WORK_DIR"
 
-    local -i LITE_CLIENT
-    declare -i VERBOSE #< Mot local - needed by onExit().
-    parseArgs "$@"
-
-    echo "Creating installer in $WORK_DIR (will be deleted on success)."
+    if [ KEEP_WORK_DIR = 1 ]; then
+        local -r WORK_DIR_NOTE="(ATTENTION: will NOT be deleted)"
+    else
+        local -r WORK_DIR_NOTE="(will be deleted on success)"
+    fi
+    echo "Creating distribution in $WORK_DIR $WORK_DIR_NOTE."
 
     if [ $VERBOSE = 0 ]; then
         redirectOutputToLog
     fi
 
-    createArmInstaller
+    buildDistribution
 
-    rm -rf "$WORK_DIR"
+    if [ KEEP_WORK_DIR = 0 ]; then
+        rm -rf "$WORK_DIR"
+    fi
 }
 
 main "$@"
